@@ -1,10 +1,11 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
+    id 'kotlin-parcelize'
 }
-apply plugin: 'kotlin-android-extensions'
 
 android {
+    namespace 'com.emm.retrofit'
     compileSdk 34
 
     defaultConfig {
@@ -24,13 +25,17 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
+
     kotlinOptions {
-        jvmTarget = '1.8'
+        jvmTarget = '17'
     }
-    namespace 'com.emm.retrofit'
+
+    buildFeatures {
+        viewBinding true
+    }
 }
 
 dependencies {
